+++ lib/stealth/generators/builder/bot/controllers/goodbyes_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class GoodbyesController < BotController
 
   def say_goodbye