+++ lib/stealth/generators/builder/bot/controllers/interrupts_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class InterruptsController < BotController
+
+  def say_interrupted
+    do_nothing
+  end
+
+end