+++ lib/stealth/logger.rb
@@ -27,7 +27,7 @@ module Stealth
 
     def self.log(topic:, message:)
       unless ENV['STEALTH_ENV'] == 'test'
-        puts "#{print_topic(topic)} #{message}"
+        puts "TID-#{Stealth.tid} #{print_topic(topic)} #{message}"
       end
     end
 
@@ -39,6 +39,8 @@ module Stealth
                 :green
               when :previous_session, :back_to_session
                 :yellow
+              when :interrupt
+                :magenta
               when :facebook, :twilio, :bandwidth
                 :blue
               when :smooch